Some updates to file structure and processing the if statement.
