+++ conditionals/conditionals.go
@@ -3,9 +3,23 @@ package conditionals
 import (
 	"../functions"
 	"../structs"
-	//"errors"
-	"fmt"
 )
+
+func ExecInput(expression structs.List, symbols *map[string]rune,
+	functionTable *map[string]structs.Function, bindings *map[string]string) (interface{}, error) {
+
+	// Check for built-in commands
+	switch (*symbols)[expression.Head.Data.(string)] {
+	case 'c':
+		value, err := EvalConditional(expression, symbols, functionTable, bindings)
+		return value, err
+	case 'f':
+		value, err := functions.ExecFunction(expression, symbols, functionTable, bindings)
+		return value, err
+	default:
+		return expression, nil
+	}
+}
 
 func EvalConditional(expression structs.List, symbols *map[string]rune,
 	functionTable *map[string]structs.Function, bindings *map[string]string) (interface{}, error) {
@@ -24,7 +38,7 @@ func if_statement(expression structs.List, symbols *map[string]rune,
 
 	e = e.Next()
 	c := e.Data.(structs.List)
-	value, err := functions.ExecFunction(c, symbols, functionTable, bindings)
+	value, err := ExecInput(c, symbols, functionTable, bindings)
 	if err != nil {
 		return nil, err
 	}
@@ -36,14 +50,14 @@ func if_statement(expression structs.List, symbols *map[string]rune,
 
 	if value == true {
 		// need to check if t is a function or a value
-		if retVal, err := main.ExecInput(t, symbols, functionTable, bindings); err != nil {
+		if retVal, err := ExecInput(t, symbols, functionTable, bindings); err != nil {
 			return nil, err
 		} else {
 			return retVal, nil
 		}
 	} else {
 		// need to check if t is a function or a value
-		if retVal, err := main.ExecInput(f, symbols, functionTable, bindings); err != nil {
+		if retVal, err := ExecInput(f, symbols, functionTable, bindings); err != nil {
 			return nil, err
 		} else {
 			return retVal, nil
