+++ main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"./conditionals"
-	"./functions"
 	"./parse"
 	"./structs"
 	"bufio"
@@ -52,14 +51,14 @@ func main() {
 		var s_expressions structs.List
 		var value interface{}
 		args := strings.Split(input, " ")
-		_, s_expressions = transliterate(s_expressions, args)
+		_, s_expressions = parse.Transliterate(s_expressions, args)
 
 		if err = parse.Parse(input); err != nil {
 			fmt.Fprintln(os.Stderr, err)
 		}
 
 		// Handle the execution of the input.
-		if value, err = execInput(s_expressions, &symbols, &functions, &bindings); err != nil {
+		if value, err = conditionals.ExecInput(s_expressions, &symbols, &functions, &bindings); err != nil {
 			fmt.Fprintln(os.Stderr, err)
 		}
 		if value != nil {
@@ -80,77 +79,5 @@ func main() {
 				fmt.Println(")")
 			}
 		}
-	}
-}
-
-func transliterate(list structs.List, args []string) (int, structs.List) {
-	expressionCount := 0
-	catchUpIndex := 0
-	currIndex := 0
-
-	for index, arg := range args {
-		currIndex = index
-		if catchUpIndex > 0 {
-			catchUpIndex--
-			continue
-		}
-		if strings.Contains(arg, ")\n") {
-			if strings.Contains(arg, "(") {
-				arg = strings.Trim(arg, "()\n")
-				list.PushBack(arg)
-			} else {
-				arg = strings.Trim(arg, ")\n")
-				list.PushBack(arg)
-			}
-			break
-		} else if strings.Contains(arg, ")") {
-			if strings.Contains(arg, "(") {
-				arg = strings.Trim(arg, "()\n")
-				list.PushBack(arg)
-			} else {
-				arg = strings.TrimRight(arg, ")")
-				list.PushBack(arg)
-			}
-			break
-		} else if strings.Contains(arg, "'(") && expressionCount == 0 { // beginning
-			list.PushBack("'")
-			list.PushBack(arg[2:])
-			expressionCount++
-		} else if strings.Contains(arg, "(") && expressionCount == 0 { // beginning
-			list.PushBack(arg[1:])
-			expressionCount++
-		} else if strings.Contains(arg, "'(") && expressionCount > 0 {
-			var newIndex int
-			var innerList structs.List
-			list.PushBack("'")
-			newIndex, innerList = transliterate(innerList, args[index:])
-			catchUpIndex = newIndex
-			list.PushBack(innerList)
-		} else if strings.Contains(arg, "(") && expressionCount > 0 {
-			var newIndex int
-			var innerList structs.List
-			newIndex, innerList = transliterate(innerList, args[index:])
-			catchUpIndex = newIndex
-			list.PushBack(innerList)
-		} else {
-			list.PushBack(strings.TrimRight(arg, ")"))
-		}
-	}
-	return currIndex, list
-}
-
-func ExecInput(expression structs.List, symbols *map[string]rune,
-	functionTable *map[string]structs.Function, bindings *map[string]string) (interface{}, error) {
-
-	// Check for built-in commands
-	switch (*symbols)[expression.Head.Data.(string)] {
-	case 'c':
-		value, err := conditionals.EvalConditional(expression, symbols, functionTable, bindings)
-		return value, err
-	case 'f':
-		value, err := functions.ExecFunction(expression, symbols, functionTable, bindings)
-		return value, err
-	default:
-		return expression, nil
 	}
 }
